Enemy can die
Added an enemy hit method to ensure the enemy takes damage from
projectiles and a die method to ensure that when the enemy loses enough
health the collider is disabled and the death boolean is made true. If
the enemy is dead the tower will stop attacking it and go to the
nearest enemy that is alive.